scraper/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
beautifulsoup4==4.6.0
feedparser==5.2.1
lxml==3.8.0
peewee==2.10.1
prometheus-client==0.0.19
pymysql==0.7.11
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0               # via python-dateutil
